% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presets.R
\name{available_journal_presets}
\alias{available_journal_presets}
\title{List available journal presets}
\usage{
available_journal_presets()
}
\value{
A character vector of supported journal preset names.
}
\description{
List available journal presets
}
