% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{available_lab_palettes}
\alias{available_lab_palettes}
\title{List available discrete palettes}
\usage{
available_lab_palettes()
}
\value{
A character vector of available discrete palette names.
}
\description{
List available discrete palettes
}
