% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presets.R
\name{available_output_presets}
\alias{available_output_presets}
\title{List available output presets}
\usage{
available_output_presets()
}
\value{
A character vector of supported save preset names.
}
\description{
List available output presets
}
