% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presets.R
\name{get_fig_size_cm}
\alias{get_fig_size_cm}
\title{Compute figure width and height from a panel specification}
\usage{
get_fig_size_cm(spec, ncol = 1, nrow = 1)
}
\arguments{
\item{spec}{Canonical panel specification name.}

\item{ncol}{Number of panel columns.}

\item{nrow}{Number of panel rows.}
}
\value{
A named list with width and height in cm.
}
\description{
Compute figure width and height from a panel specification
}
