% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presets.R
\name{journal_preset}
\alias{journal_preset}
\title{Get a named journal preset}
\usage{
journal_preset(name)
}
\arguments{
\item{name}{Journal preset name.}
}
\value{
A named list of journal-specific defaults.
}
\description{
Get a named journal preset
}
