% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{lab_gradient_palette}
\alias{lab_gradient_palette}
\title{Return a lab gradient palette}
\usage{
lab_gradient_palette(n = 256, palette = "blue_red", reverse = FALSE)
}
\arguments{
\item{n}{Number of colors required.}

\item{palette}{Gradient palette name.}

\item{reverse}{Whether to reverse the palette.}
}
\value{
A character vector of hex colors.
}
\description{
Return a lab gradient palette
}
