% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{lab_palette}
\alias{lab_palette}
\title{Return a lab discrete palette}
\usage{
lab_palette(n, palette = "main", reverse = FALSE, alpha = 1)
}
\arguments{
\item{n}{Number of colors required.}

\item{palette}{Palette name.}

\item{reverse}{Whether to reverse the palette.}

\item{alpha}{Opacity multiplier between \code{0} and \code{1}.}
}
\value{
A character vector of hex colors.
}
\description{
Return a lab discrete palette
}
