% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presets.R
\name{panel_size}
\alias{panel_size}
\title{Get a named panel size preset}
\usage{
panel_size(name)
}
\arguments{
\item{name}{Preset name.}
}
\value{
A named list describing the preset dimensions in cm.
}
\description{
Get a named panel size preset
}
