% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_color_lab}
\alias{scale_color_lab}
\title{Discrete lab color scale}
\usage{
scale_color_lab(
  palette = "main",
  reverse = FALSE,
  alpha = 1,
  na.value = "#BFBFBF",
  guide = ggplot2::guide_legend(),
  ...
)
}
\arguments{
\item{palette}{Palette name.}

\item{reverse}{Whether to reverse the palette.}

\item{alpha}{Opacity multiplier between \code{0} and \code{1}.}

\item{na.value}{Fallback color for missing values.}

\item{guide}{Legend guide. Use \code{"none"} to suppress the guide.}

\item{...}{Passed to \code{ggplot2::discrete_scale()}.}
}
\value{
A ggplot scale.
}
\description{
Discrete lab color scale
}
