% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_fill_lab}
\alias{scale_fill_lab}
\title{Discrete lab fill scale}
\usage{
scale_fill_lab(
  palette = "main",
  reverse = FALSE,
  alpha = 1,
  na.value = "#BFBFBF",
  ...
)
}
\arguments{
\item{palette}{Palette name.}

\item{reverse}{Whether to reverse the palette.}

\item{alpha}{Opacity multiplier between \code{0} and \code{1}.}

\item{na.value}{Fallback color for missing values.}

\item{...}{Passed to \code{ggplot2::discrete_scale()}.}
}
\value{
A ggplot scale.
}
\description{
Discrete lab fill scale
}
