% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{available_group_dictionaries}
\alias{available_group_dictionaries}
\title{List available semantic color dictionaries}
\usage{
available_group_dictionaries()
}
\value{
A character vector of available dictionary names.
}
\description{
List available semantic color dictionaries
}
