% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_lab}
\alias{layout_lab}
\title{Compose multiple plots with a lab-standard patchwork layout}
\usage{
layout_lab(..., ncol = NULL, nrow = NULL, guides = "collect")
}
\arguments{
\item{...}{ggplot objects or a single list of ggplot objects.}

\item{ncol}{Number of columns in the composed layout.}

\item{nrow}{Number of rows in the composed layout.}

\item{guides}{Guide collection strategy passed to \code{patchwork::wrap_plots()}.}
}
\value{
A patchwork composition.
}
\description{
Compose multiple plots with a lab-standard patchwork layout
}
